Add rename-file-or-dir
This op is used to move a file or a directory of files.  If any clojure
files are affected by the move any dependent namespaces are updated.

Unlikely tools.namespace's move-ns this op also handles vectors written
in prefix notation as well as fully qualified references in the file
body itself referencing vars from the old namespace.